pyproject.toml
[project]
name = "bicycle-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "ipykernel>=6.30.1",
    "llama-index>=0.12.52",
    "openai>=1.101.0",
    "pandas>=2.3.2",
    "psycopg[binary]>=3.2.9",
    "pydantic-ai-slim[openai]>=0.8.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = ["black>=25.1.0", "isort>=6.0.1", "pytest>=8.4.1"]
